mcv/requirements.txt
Django==4.1.5
django_xhtml2pdf==0.0.4
sorl_thumbnail==12.9.0
